explicit cast

+++ physarumgen/src/world.cpp
@@ -27,7 +27,7 @@ bool c_world::spawn_agents(float world_percentage)
     for(int i=0; i<num_agents; i++)
     {
         s_vec2 tmp_pos = m_world_sampler.get_next_position();
-        m_agents.emplace_back(tmp_pos.x, tmp_pos.y, m_world_sampler.get_angle());
+        m_agents.emplace_back((float)tmp_pos.x, (float)tmp_pos.y, m_world_sampler.get_angle());
     }
 
     std::cout << "Spawned " << m_agents.size() << " agents" << std::endl;
